Rename examples file
