+++ packages/docs-app/src/examples/core-examples/hotkeysTargetExample.tsx
